python-ai/requirements.txt
flask==3.0.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==1.5.3
joblib==1.3.2
pyyaml==6.0.1
requests==2.31.0
scipy==1.11.4
matplotlib==3.8.2